+++ source/xmldocument.cpp
@@ -136,7 +136,7 @@ namespace CrissCross
 				
 				size_t fileLen = _file->Length();
 				char *buffer = new char[fileLen + 1];
-				_file->Read ( buffer, fileLen + 1, 0, fileLen );
+				_file->Read ( buffer, (int)fileLen + 1, 0, (int)fileLen );
 				
 				char *data = new char[fileLen + 1];
 				char const *p = buffer;
